requirements.txt
python-telegram-bot>=20.0
groq>=0.4.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
yfinance>=0.2.0
textblob>=0.17.0
websocket-client>=1.6.0
pandas>=1.5.0
numpy>=1.24.0
